+++ Shared/Realm.Shared/Extensions/RealmResultsCollectionChanged.cs
@@ -28,6 +28,23 @@ namespace Realms
     public static class RealmResultsCollectionChanged
     {
         /// <summary>
+        /// A convenience method that casts <c>IOrderedQueryable{T}</c> to <see cref="RealmResults{T}"/> which implements INotifyCollectionChanged.
+        /// </summary>
+        /// <param name="results">The <see cref="RealmResults{T}" /> to observe for changes.</param>
+        /// <typeparam name="T">Type of the RealmObject in the results.</typeparam>
+        /// <seealso cref="RealmResults{T}.SubscribeForNotifications(RealmResults{T}.NotificationCallbackDelegate)"/>
+        /// <returns>The <see cref="RealmResults{T}"/> instance cast to <see cref="INotifyCollectionChanged"/>.</returns>
+        public static INotifyCollectionChanged ToNotifyCollectionChanged<T>(this IOrderedQueryable<T> results) where T : RealmObject
+        {
+            if (!(results is RealmResults<T>))
+            {
+                throw new ArgumentException($"{nameof(results)} must be an instance of RealmResults<{typeof(T).Name}>", nameof(results));
+            }
+
+            return (RealmResults<T>)results;
+        }
+
+        /// <summary>
         /// Wraps a <see cref="RealmResults{T}" /> in an implementation of <see cref="INotifyCollectionChanged" /> so that it may be used in MVVM databinding.
         /// </summary>
         /// <param name="results">The <see cref="RealmResults{T}" /> to observe for changes.</param>
@@ -35,6 +52,7 @@ namespace Realms
         /// <typeparam name="T">Type of the RealmObject in the results.</typeparam>
         /// <returns>An <see cref="ObservableCollection{T}" />-like object useful for MVVM databinding.</returns>
         /// <seealso cref="RealmResults{T}.SubscribeForNotifications(RealmResults{T}.NotificationCallbackDelegate)"/>
+        [Obsolete("RealmResults now implements INotifyPropertyChanged, so this method is no longer needed.")]
         public static INotifyCollectionChanged ToNotifyCollectionChanged<T>(this IOrderedQueryable<T> results, Action<Exception> errorCallback) where T : RealmObject
         {
             return ToNotifyCollectionChanged(results, errorCallback, coalesceMultipleChangesIntoReset: false);
@@ -52,6 +70,7 @@ namespace Realms
         /// <typeparam name="T">Type of the RealmObject in the results.</typeparam>
         /// <returns>An <see cref="ObservableCollection{T}" />-like object useful for MVVM databinding.</returns>
         /// <seealso cref="RealmResults{T}.SubscribeForNotifications(RealmResults{T}.NotificationCallbackDelegate)"/>
+        [Obsolete("RealmResults now implements INotifyPropertyChanged, so this method is no longer needed.")]
         public static INotifyCollectionChanged ToNotifyCollectionChanged<T>(this IOrderedQueryable<T> results, Action<Exception> errorCallback, bool coalesceMultipleChangesIntoReset) where T : RealmObject
         {
             if (results == null)
@@ -69,102 +88,7 @@ namespace Realms
                 throw new ArgumentNullException(nameof(errorCallback));
             }
 
-            return new ReadOnlyObservableCollection<T>(new Adapter<T>((RealmResults<T>)results, errorCallback, coalesceMultipleChangesIntoReset));
-        }
-
-        public sealed class Adapter<T> : ObservableCollection<T> where T : RealmObject
-        {
-            private readonly RealmResults<T> _results;
-            private readonly IDisposable _token;
-            private readonly Action<Exception> _errorCallback;
-            private readonly bool _coalesceMultipleChangesIntoReset;
-
-            private bool _suspendNotifications;
-
-            public Adapter(RealmResults<T> results, Action<Exception> errorCallback, bool coalesceMultipleChangesIntoReset) : base(results)
-            {
-                _results = results;
-                _errorCallback = errorCallback;
-                _coalesceMultipleChangesIntoReset = coalesceMultipleChangesIntoReset;
-
-                _token = results.SubscribeForNotifications(OnChange);
-                Debug.Assert(_token != null, "Subscription token must not be null.");
-            }
-
-            ~Adapter()
-            {
-                _token.Dispose();
-            }
-
-            private void OnChange(RealmResults<T> sender, RealmResults<T>.ChangeSet change, Exception error)
-            {
-                if (error != null)
-                {
-                    _errorCallback(error);
-                }
-                else if (change != null)
-                {
-                    _suspendNotifications = _coalesceMultipleChangesIntoReset && change.InsertedIndices.Length + change.DeletedIndices.Length > 1;
-
-                    foreach (var removed in change.DeletedIndices.Reverse())
-                    {
-                        // the row has been deleted, we need to recreate the adapter's contents
-                        if (!this[removed].ObjectHandle.IsValid)
-                        {
-                            Recreate();
-                            return;
-                        }
-
-                        RemoveAt(removed);
-                    }
-
-                    foreach (var added in change.InsertedIndices)
-                    {
-                        InsertItem(added, _results[added]);
-                    }
-
-                    if (_suspendNotifications)
-                    {
-                        RaiseReset();
-                        _suspendNotifications = false;
-                    }
-                }
-            }
-
-            private void Recreate()
-            {
-                _suspendNotifications = true;
-                this.Clear();
-                foreach (var item in _results)
-                {
-                    this.Add(item);
-                }
-
-                _suspendNotifications = false;
-                RaiseReset();
-            }
-
-            private void RaiseReset()
-            {
-                base.OnCollectionChanged(new NotifyCollectionChangedEventArgs(NotifyCollectionChangedAction.Reset));
-                base.OnPropertyChanged(new PropertyChangedEventArgs(nameof(Count)));
-            }
-
-            protected override void OnCollectionChanged(NotifyCollectionChangedEventArgs e)
-            {
-                if (!_suspendNotifications)
-                {
-                    base.OnCollectionChanged(e);
-                }
-            }
-
-            protected override void OnPropertyChanged(PropertyChangedEventArgs e)
-            {
-                if (!_suspendNotifications)
-                {
-                    base.OnPropertyChanged(e);
-                }
-            }
+            return (RealmResults<T>)results;
         }
     }
 }